requirements.txt
boto3
asn1crypto
certifi
cffi
charset-normalizer
cryptography
idna
oscrypto
pycparser
pycryptodomex
PyJWT
pyOpenSSL
pytz
requests
six
urllib3
snowflake-connector-python
cffi
psycopg2-binary